requirements.txt
Flask
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Werkzeug==2.3.7